requirements.txt
numpy==2.1.3
pandas==2.2.3
pyinstaller==6.11.1
pykrx==1.0.48
ccxt==4.4.50
statsmodels==0.14.4
PyQt6==6.8.0
PySide6==6.8.1.1
matplotlib==3.10.0
torch==2.6.0+cu126
torchaudio==2.6.0+cu126
torchvision==0.21.0+cu126
yfinance==0.2.52
jupyterlab==4.3.4
jupyterlab-execute-time==3.2.0
seaborn==0.13.2
scikit-learn==1.6.1
openai==1.61.0
pyarrow==19.0.0
gspread==6.1.4
google-auth==2.38.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.161.0
openpyxl==3.1.5
fastapi==0.115.8
bcrypt==4.2.1
gunicorn==23.0.0
httpie==3.2.4
httpx==0.28.1
Jinja2==3.1.5
pydantic==2.10.6
pydantic-core==2.27.2
pydantic-extra-types==2.10.2
pydantic-settings==2.7.1
pytest==8.3.4
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.20
requests==2.32.3
starlette==0.45.3
tabulate==0.9.0
uvicorn==0.34.0
uvloop==0.21.0; sys_platform != "win32"
plotly==6.0.0
anywidget==0.9.14
xgboost==2.1.4

slicer==0.0.8
numba==0.61.0
cloudpickle==3.1.1

imbalanced-learn==0.13.0

pandas-market-calendars==4.6.1

